+++ controllers/subjectController.js
@@ -2,6 +2,9 @@ const Subject = require("../models/subject");
 
 exports.createSubject = async function (req, res) {
   try {
+    if (!req.user.isLoggedIn) {
+      return res.status(400).json({ message: "Please Log in" });
+    }
     const subject = new Subject(req.body);
     await subject.save();
     res.json(subject);
@@ -12,6 +15,9 @@ exports.createSubject = async function (req, res) {
 
 exports.showSubject = async function (req, res) {
   try {
+    if (!req.user.isLoggedIn) {
+      return res.status(400).json({ message: "Please Log in" });
+    }
     const subject = await Subject.findOne({ _id: req.params.id });
     if (!subject) {
       return res.status(404).json({ message: "Subject not found" });
@@ -24,6 +30,9 @@ exports.showSubject = async function (req, res) {
 
 exports.showAllSubjects = async function (req, res) {
   try {
+    if (!req.user.isLoggedIn) {
+      return res.status(400).json({ message: "Please Log in" });
+    }
     const subjects = await Subject.find({});
     res.json(subjects);
   } catch (error) {
@@ -32,6 +41,9 @@ exports.showAllSubjects = async function (req, res) {
 };
 exports.updateSubject = async function (req, res) {
   try {
+    if (!req.user.isLoggedIn) {
+      return res.status(400).json({ message: "Please Log in" });
+    }
     const updates = Object.keys(req.body);
     const subject = await Subject.findOne({ _id: req.params.id });
     if (!subject) {
@@ -47,6 +59,9 @@ exports.updateSubject = async function (req, res) {
 
 exports.deleteSubject = async function (req, res) {
   try {
+    if (!req.user.isLoggedIn) {
+      return res.status(400).json({ message: "Please Log in" });
+    }
     const subject = await Subject.findOneAndRemove({ _id: req.params.id });
     if (!subject) {
       return res.status(404).json({ message: "Subject not found" });
